Created auth.py, db.py, log.py, and moved related code from app.py created blueprints folder and created home.py and pos.py and moved related api endpoints from app.py moved toJson to utils.com.common removed commented codes
