+++ src/mtool/api/UnitTests/POS_Operations/test_pos.py
@@ -50,7 +50,7 @@ DAGENT_URL = 'http://' + ip + ':3000'
 INFLUXDB_URL = 'http://0.0.0.0:8086/write?db=poseidon&rp=autogen'
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.connection_factory.get_current_user",
+@mock.patch("rest.db.connection_factory.get_current_user",
             return_value="test", autospec=True)
 def test_is_ibofos_running(mock_get_current_user, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
@@ -74,9 +74,9 @@ def test_is_ibofos_running(mock_get_current_user, **kwargs):
 
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.connection_factory.get_current_user",
+@mock.patch("rest.db.connection_factory.get_current_user",
             return_value="test", autospec=True)
-@mock.patch("rest.app.fetch_system_state", return_value={'lastSuccessTime':1}, autospec=True)
+@mock.patch("rest.blueprints.pos.fetch_system_state", return_value={'lastSuccessTime':1}, autospec=True)
 def test_is_ibofos_running_failure(mock_get_current_user, mock_fetch_system_state,  **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].get(DAGENT_URL + '/api/dagent/v1/heartbeat',
@@ -91,7 +91,7 @@ def test_is_ibofos_running_failure(mock_get_current_user, mock_fetch_system_stat
 """
 @requests_mock.Mocker(kw="mock")
 @mock.patch("rest.app.get_rebuilding_status",return_value=(), autospec=True)
-@mock.patch("rest.app.connection_factory.get_current_user", return_value="test", autospec=True)
+@mock.patch("rest.db.connection_factory.get_current_user", return_value="test", autospec=True)
 def test_is_ibofos_runningi_failure(mock_get_current_user, mock_get_rebuilding_status, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].get(DAGENT_URL + '/api/dagent/v1/heartbeat',
@@ -104,9 +104,9 @@ def test_is_ibofos_runningi_failure(mock_get_current_user, mock_get_rebuilding_s
 """
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.get_ibof_os_status",return_value=False, autospec=True)
-@mock.patch("rest.app.is_ibofos_running",return_value=False, autospec=True)
-@mock.patch("rest.app.connection_factory.get_current_user",return_value="test", autospec=True)
+@mock.patch("rest.blueprints.pos.get_ibof_os_status",return_value=False, autospec=True)
+@mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=False, autospec=True)
+@mock.patch("rest.db.connection_factory.get_current_user",return_value="test", autospec=True)
 def test_start_ibofos(mock_get_current_user,mock_get_ibof_os_status,mock_is_ibofos_running, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].post(DAGENT_URL + '/api/ibofos/v1/system',
@@ -130,7 +130,7 @@ def test_start_ibofos(mock_get_current_user,mock_get_ibof_os_status,mock_is_ibof
 
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.connection_factory.get_current_user",return_value="test", autospec=True)
+@mock.patch("rest.db.connection_factory.get_current_user",return_value="test", autospec=True)
 def test_start_ibofos_dagent_ibofos(mock_get_current_user,**kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].post(DAGENT_URL + '/api/ibofos/v1/system',
@@ -140,7 +140,7 @@ def test_start_ibofos_dagent_ibofos(mock_get_current_user,**kwargs):
     assert response.status_code == 200
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.connection_factory.get_current_user",return_value="test", autospec=True)
+@mock.patch("rest.db.connection_factory.get_current_user",return_value="test", autospec=True)
 def test_get_pos_info(mock_get_current_user, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].get(DAGENT_URL + '/api/ibofos/v1/system',
@@ -153,7 +153,7 @@ def test_get_pos_info(mock_get_current_user, **kwargs):
     assert response.status_code == 200
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.connection_factory.get_current_user",return_value="test", autospec=True)
+@mock.patch("rest.db.connection_factory.get_current_user",return_value="test", autospec=True)
 def test_get_pos_info_err(mock_get_current_user, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].get(DAGENT_URL + '/api/ibofos/v1/system', exc=requests.exceptions.ConnectTimeout)
@@ -165,9 +165,9 @@ def test_get_pos_info_err(mock_get_current_user, **kwargs):
     assert response.status_code == 500
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.get_ibof_os_status",return_value=False, autospec=True)
-@mock.patch("rest.app.connection_factory.get_current_user",return_value="test", autospec=True)
-@mock.patch("rest.app.is_ibofos_running",return_value=True, autospec=True)
+@mock.patch("rest.blueprints.pos.get_ibof_os_status",return_value=False, autospec=True)
+@mock.patch("rest.db.connection_factory.get_current_user",return_value="test", autospec=True)
+@mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=True, autospec=True)
 def test_start_ibofos_already_running_failure(mock_get_current_user, mock_get_ibof_os_status, mock_is_ibofos_running, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].post(DAGENT_URL + '/api/ibofos/v1/system',
@@ -182,10 +182,10 @@ def test_start_ibofos_already_running_failure(mock_get_current_user, mock_get_ib
 
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.connection_factory.get_current_user",
+@mock.patch("rest.db.connection_factory.get_current_user",
             return_value="test", autospec=True)
-@mock.patch("rest.app.is_ibofos_running",return_value=False, autospec=True)
-@mock.patch("rest.app.get_ibof_os_status",return_value=False, autospec=True)
+@mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=False, autospec=True)
+@mock.patch("rest.blueprints.pos.get_ibof_os_status",return_value=False, autospec=True)
 def test_start_ibofos_failure(mock_get_current_user, mock_get_ibof_os_status,mock_is_ibofos_running, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].post(DAGENT_URL + '/api/ibofos/v1/system',
@@ -207,9 +207,9 @@ def test_start_ibofos_failure(mock_get_current_user, mock_get_ibof_os_status,moc
 
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.get_ibof_os_status",return_value=False, autospec=True)
-@mock.patch("rest.app.connection_factory.get_current_user",return_value="test", autospec=True)
-@mock.patch("rest.app.is_ibofos_running",return_value=False, autospec=True)
+@mock.patch("rest.blueprints.pos.get_ibof_os_status",return_value=False, autospec=True)
+@mock.patch("rest.db.connection_factory.get_current_user",return_value="test", autospec=True)
+@mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=False, autospec=True)
 def test_start_ibofos_already_running(mock_get_current_user,mock_get_ibof_os_statusi, mock_is_ibofos_running,  **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].post(DAGENT_URL + '/api/ibofos/v1/system',
@@ -227,9 +227,9 @@ def test_start_ibofos_already_running(mock_get_current_user,mock_get_ibof_os_sta
 
 """
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.connection_factory.get_current_user",
+@mock.patch("rest.db.connection_factory.get_current_user",
             return_value="test", autospec=True)
-@mock.patch("rest.app.is_ibofos_running",return_value=False, autospec=True)
+@mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=False, autospec=True)
 def test_stop_ibofos_already_running(mock_get_current_user,mock_is_ibofos_running, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].get(DAGENT_URL + '/api/ibofos/v1/system/mount',
@@ -247,11 +247,11 @@ def test_stop_ibofos_already_running(mock_get_current_user,mock_is_ibofos_runnin
 
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.connection_factory.get_current_user",
+@mock.patch("rest.db.connection_factory.get_current_user",
             return_value="test", autospec=True)
-@mock.patch("rest.app.IBOF_OS_Running.Is_Ibof_Os_Running_Flag",
+@mock.patch("rest.blueprints.pos.IBOF_OS_Running.Is_Ibof_Os_Running_Flag",
             return_value=True, autospec=True)
-@mock.patch("rest.app.is_ibofos_running",return_value=False, autospec=True)
+@mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=False, autospec=True)
 def test_stop_ibofos(mock_get_current_user,mock_Is_Ibof_Os_Running_Flag, mock_is_ibofos_running, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].delete(DAGENT_URL + '/api/ibofos/v1/system/mount',
@@ -272,11 +272,11 @@ def test_stop_ibofos(mock_get_current_user,mock_Is_Ibof_Os_Running_Flag, mock_is
 
 
 @requests_mock.Mocker(kw="mock")
-@mock.patch("rest.app.connection_factory.get_current_user",
+@mock.patch("rest.db.connection_factory.get_current_user",
             return_value="test", autospec=True)
-@mock.patch("rest.app.IBOF_OS_Running.Is_Ibof_Os_Running_Flag",
+@mock.patch("rest.blueprints.pos.IBOF_OS_Running.Is_Ibof_Os_Running_Flag",
             return_value=True, autospec=True)
-@mock.patch("rest.app.is_ibofos_running",return_value=False, autospec=True)
+@mock.patch("rest.blueprints.pos.is_ibofos_running",return_value=False, autospec=True)
 def test_stop_ibofos_failure(mock_get_current_user,mock_Is_Ibof_Os_Running_Flag, mock_is_ibofos_running, **kwargs):
     kwargs["mock"].post(INFLUXDB_URL, text='Success', status_code=204)
     kwargs["mock"].delete(DAGENT_URL + '/api/ibofos/v1/system',
